apiservice/requirements.txt
environs

fastapi~=0.100.0
SQLAlchemy~=2.0.19
python-dotenv~=1.0.0
pydantic~=2.0.3
alembic~=1.11.1
redis~=4.6.0